refactor(multipart): move the streaming traversal into PartScanner
MultipartUtil held two unrelated APIs: split(), which buffers every part into a
List, and forEachPart(), which streams parts to a handler that decides from each
header whether the body is needed at all. The streaming one had brought four
nested types with it, so a class named *Util ended up owning a callback
protocol - callers wrote MultipartUtil.PartHandler and caught
MultipartUtil.PartTooLargeException, where the qualifier says nothing.

Move forEachPart and its internals to PartScanner, and give the exception its
own file: the project keeps exceptions top-level almost without exception (47
own-file against 4 public nested). PartAction and PartHandler stay nested, now
in the class that owns them - public nested enums and single-method callbacks
tied to one method are the established convention here.

MultipartUtil drops from 259 to 114 lines and keeps isMultipart and split.
The tests follow the split: MultipartUtilTest keeps the split contract,
PartScannerTest takes the traversal.
